requirements.txt
httpx>=0.27.0
eth-account>=0.13.0
eth-utils>=5.0.0
msgpack>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-timeout>=2.2.0
pytest-order>=1.2.0
allure-pytest>=2.13.0
